fix: lock condition

+++ postgres/schemas/models.py
@@ -365,7 +365,7 @@ def update_dictionary(engine, theme_keywords):
 
 
 def empty_tables(session = None, stop_word = True):
-    if(os.environ.get("POSTGRES_HOST") == "postgres_db" or os.environ.get("POSTGRES_HOST") == "localhost"):
+    if( (os.environ.get("POSTGRES_HOST") == "postgres_db" or os.environ.get("POSTGRES_HOST") == "localhost") and os.environ.get("ENV") != "prod"):
         logging.warning("""Doing: Empty table Stop_Word / Keywords""")
         if stop_word:
             session.query(Stop_Word).delete()
@@ -377,16 +377,20 @@ def empty_tables(session = None, stop_word = True):
 
 def drop_tables(conn = None):
     
-    if(os.environ.get("POSTGRES_HOST") == "postgres_db" or os.environ.get("POSTGRES_HOST") == "localhost"):
+    if( (os.environ.get("POSTGRES_HOST") == "postgres_db" or os.environ.get("POSTGRES_HOST") == "localhost") and os.environ.get("ENV") != "prod"):
         logging.warning("""Drop table keyword / Program_Metadata / Channel_Metadata in the PostgreSQL database""")
         try:
             if conn is None :
                 engine = connect_to_db()
             else:
                 engine = conn
+            logging.info(f"Drop all {Keywords.__tablename__}")
             Base.metadata.drop_all(bind=engine, tables=[Keywords.__table__])
+            logging.info(f"Drop all {Channel_Metadata.__tablename__}")
             Base.metadata.drop_all(bind=engine, tables=[Channel_Metadata.__table__])
+            logging.info(f"Drop all {Program_Metadata.__tablename__}")
             Base.metadata.drop_all(bind=engine, tables=[Program_Metadata.__table__])
+            logging.info(f"Drop all {Stop_Word.__tablename__}")
             Base.metadata.drop_all(bind=engine, tables=[Stop_Word.__table__])
 
             logging.info(f"Table keyword / Program_Metadata / Channel_Metadata deletion done")
